requirements.txt
transformers>=4.30.0
torch>=2.0.0
yfinance>=0.2.18
praw>=7.7.0
newsapi-python>=0.2.7
feedparser>=6.0.10
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
plotly>=5.15.0
streamlit>=1.24.0
requests>=2.31.0
pytest>=7.0.0
python-dotenv>=1.0.0
